+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- **[FIX]** Fix back button alignment in `SearchForm` autocomplete sections.
 [...]
 
 # v34.1.0 (27/05/2020)