+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- **[NEW]** Add `disabled` prop to `Stepper`
 - **[FIX]** Fix back button alignment in `SearchForm` autocomplete sections.
 - **[UPDATE]** Only show the `SearchForm` switch button when at least FROM or TO is selected.
 [...]