+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Unreleased
 
-- **[NEW]** Add `disabled` prop to `Stepper`
-- **[FIX]** Fix back button alignment in `SearchForm` autocomplete sections.
-- **[UPDATE]** Only show the `SearchForm` switch button when at least FROM or TO is selected.
 [...]
+
+# v34.2.0 (27/05/2020)
+
+- **[NEW]** Add `disabled` prop to `Stepper`
+- **[UPDATE]** Only show the `SearchForm` switch button when at least FROM or TO is selected.
 - **[FIX]** Remove hover color & pointer when `disabled` is passed to `ItemCheckbox`/`ItemRadio`
+- **[FIX]** Fix back button alignment in `SearchForm` autocomplete sections.
 
 # v34.1.0 (27/05/2020)